pyproject.toml
[tool.poetry]
name = "cag"
version = "0.1.0"
description = "Contract Automated Generator (CAG): Automated contract generation using LLMs, retrieval, and agentic workflows."
authors = ["Israel Mata <isra@example.com>", "Ricardo Alanis <ricardo@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = ">0.3.0"
langgraph = ">0.4.6"
pydantic = "^2.7.0"
faiss-cpu = "^1.8.0"
pypdf2 = "^3.0.1"
pdfplumber = "^0.10.3"
langfuse = ">=2.20.0"
python-dotenv = "^1.0.0"
scikit-learn = "^1.4.0"
numpy = "^1.26.0"
markdown2 = "^2.4.13"
xhtml2pdf = "^0.2.15"
openai = "^1.3.5"
langsmith = ">=0.3.0"
langchain-openai = ">=0.3.0"
langchain_community = ">=0.3.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
langchain_unstructured = ">0.0.10"
unstructured = {extras = ["docx", "pdf", "pptx"], version = "^0.17.2"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
langgraph-cli = {extras = ["inmem"], version = "^0.2.10"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 